backend/requirements.txt
# =====================================================
# Alcova Smart Check-in - Backend Dependencies
# =====================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
supabase==2.3.4
postgrest==0.13.2
asyncpg==0.29.0

# Tuya Integration
tuya-connector==0.1.3
# tinytuya==1.13.2  # Alternative for local control

# Communication
twilio==8.10.3
python-telegram-bot==20.7
aiosmtplib==3.0.1
emails==0.6

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
pydantic-core==2.14.5
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.15.0

# Code Quality
black==23.12.1
ruff==0.1.8
mypy==1.7.1

# Monitoring (optional)
# sentry-sdk[fastapi]==1.38.0
